feat(str): Hybrid Latin-1 / UTF-16 Str layout flip (P11.1-S1)
P11.1-S1 — internal Str representation moves from the pre-S1
byte-Str (length: u64 @8, raw bytes @16) to the post-S1 hybrid
layout (length: u32 @8, reserved: u32 @12, payload @16) with an
IS_LATIN1 flag bit added to the universal heap header flags.
Every alloc site in this commit hard-codes is_latin1=true so
ASCII / Latin-1 payloads remain byte-identical to pre-S1 — S2
will land the build-time encoding decision for non-ASCII
literals, S3 the runtime UTF-16 alloc path.

Driven by the RFC at
`.claude/rfcs/20260603-p11-1-str-unicode-internal/`. Five sign-
off decisions baked in here:

- Scheme B (Hybrid Latin-1 / UTF-16) per V8 SeqOneByteString /
  SpiderMonkey JSLinearString convention; UTF-16-only (scheme A)
  disqualified for doubling ASCII-dominant payload footprint.
- byte_capacity not stored — recomputed as
  `length × (if is_latin1 { 1 } else { 2 })` at every drop / pool
  query (mirrors V8); the saved 4 bytes become reserved @12 for
  later P11.x extensions (hash slot / inline-cache hint).
- 16-byte prefix preserved (`STR_HDR_SIZE` / `STR_DATA_OFF`) so
  all ssa_inkwell GEPs against `STR_DATA_OFF` keep resolving
  unchanged.

SSA layer:

- New `InstKind::ZExtI32ToI64(Operand)` variant + inkwell
  emitter (mirrors `ZExtBoolToI64` shape; LLVM `zext i32 to i64`).
  Lets ssa_lower widen the new u32 `length` field to the i64-
  shaped `Type::Number` value the JS spec expects.
- New encoding-aware helper `ssa_lower_str::
  load_str_or_substr_length(ctx, op, ty) -> Operand` centralises
  the Str (u32 @8 + ZExt) vs Substr (u64 @8) load fork at the
  three call sites that previously hard-coded `Load(I64, _, 8)`:
  `.length` property access, inline `s === <literal>` byte
  equality (`emit_inline_str_eq_bytes`), and `coerce_to_bool` for
  Str / Substr operands. Substr keeps its pre-S1 u64 layout —
  its code-unit index flip is queued for S5.
- god-file ssa_lower.rs: 26587 → 26581 (net -6 LOC). The three
  inlined dispatch arms collapse to single helper calls; the
  added helper lives in the ssa_lower_str sidekick (acknowledged
  P11.1 helper exception, documented in its module doc).
- Literal materialisation: `ssa_inkwell::globals::
  emit_static_str_global` rewrites the rodata Str body from
  `[u64 header, u64 length, [N x i8]]` to `[u64 header, u32
  length, u32 _pad, [N x i8]]` with the universal header carrying
  `STATIC_LITERAL | IS_LATIN1` flags. The `_pad` slot is
  zero-initialised; runtime `Load(I32, p, 8)` reads `length`
  exactly.

torajs-str crate:

- `layout.rs`: adds `STR_PAD_OFF = 12`, `STR_POOL_PAYLOAD: u32`
  (was u64), `STR_FLAG_IS_LATIN1 = 0x0002`, reshapes
  `packed_header_init(is_latin1: bool) -> u64` and
  `block_size(length: u32, is_latin1: bool) -> usize`, adds
  `byte_capacity(length, is_latin1)` helper.
- `block.rs`: `StrBlock::alloc(length: u32)` (was `u64`),
  `len() -> u32` renamed `length()`, new `is_latin1(&self)`
  query, `as_bytes_mut(byte_len: u32)`, `free_pool_aware`
  recomputes byte capacity from header flags. `init_header_and_
  length` writes packed header + length + zero-padded reserved
  slot atomically. Extern ABI for `__torajs_str_alloc_pooled`
  and `__torajs_str_alloc` keeps the FFI `u64`/`i64` `len`
  signature (so the IR-emitted call sites don't churn) but
  truncates to `u32` internally and forces `is_latin1=true`.
- All hot-path subfiles converted to the new u32 length:
  `concat.rs` / `eq.rs` / `slice.rs` / `lookup.rs` / `print.rs` /
  `json.rs` / `json_parse/*` / `literals.rs` / `to_number.rs` /
  `substr.rs` (tests only — substr's own len field stays u64) /
  `symbol.rs` / `split/ops.rs` (SplitIter ABI keeps u64, so the
  `parent_len`/`sep_len` writes cast through `as u64`) /
  `transform/{trim,pad,case,construct,replace}.rs` (`pad_output_
  len`, `substring_range`, `substr_range`, `at_resolve`,
  `repeat_out_len`, `replace_all_out_len` all reshape to u32 in
  / u32 out).
- Integration tests `tests/{perf_gate,string_ops}.rs` updated
  for the new `alloc` / `as_bytes_mut` signature.

Pre-flight:

- `cargo build --workspace --release`: 0 warn / 0 err
- `cargo fmt --check`: clean
- `cargo test -p torajs-str -p torajs-core --release`: 256
  passed / 0 failed (103 core + 148 str + 9 string_ops + 3
  perf_gate, doc-tests ignored as usual)
- `release-build.sh` + `rm -rf ~/.torajs/cache` (compiler
  fingerprint changes because `COMPILER_SOURCE_FILES` covers
  several touched files)
- Manual fixture sample (ASCII / Latin-1 payload only): 19
  str-heavy / promise-string / async / coerce / fs / template /
  switch / sub / throw / console / multi-arg-any / class-async-
  method / await-rejected-throws fixtures all bun-parity byte-
  equal. Conformance gate launched in background.

Refs:
- RFC: `.claude/rfcs/20260603-p11-1-str-unicode-internal/design.md`
- Plan state: `.claude/plan-state.md` L3a `P11.1-S1`
- Sub-step bookkeeping: tasks #1-#6 in this session

+++ crates/torajs-str/src/json.rs
@@ -32,8 +32,8 @@ use crate::layout::{STR_DATA_OFF, STR_LEN_OFF};
 const HEX: &[u8; 16] = b"0123456789abcdef";
 
 #[inline]
-fn escaped_len(s: &[u8]) -> u64 {
-    let mut out: u64 = 2; // surrounding quotes
+fn escaped_len(s: &[u8]) -> u32 {
+    let mut out: u32 = 2; // surrounding quotes
     for &c in s {
         out += match c {
             b'"' | b'\\' | b'\n' | b'\r' | b'\t' | 0x08 | 0x0c => 2,
@@ -112,7 +112,7 @@ fn write_escaped(s: &[u8], dst: &mut [u8]) {
 /// [`crate::layout`]).
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn __torajs_json_quote_str(s: *const u8) -> *mut u8 {
-    let len = unsafe { (s.add(STR_LEN_OFF) as *const u64).read() };
+    let len = unsafe { (s.add(STR_LEN_OFF) as *const u32).read() };
     let bytes = unsafe { core::slice::from_raw_parts(s.add(STR_DATA_OFF), len as usize) };
     let out_len = escaped_len(bytes);
     let mut block = StrBlock::alloc(out_len);
