+++ crates/torajs-str/tests/string_ops.rs
@@ -7,8 +7,8 @@ use torajs_str::{
 };
 
 fn make_str(payload: &[u8]) -> *mut u8 {
-    let mut b = StrBlock::alloc(payload.len() as u64);
-    let dst = unsafe { b.as_bytes_mut(payload.len() as u64) };
+    let mut b = StrBlock::alloc(payload.len() as u32);
+    let dst = unsafe { b.as_bytes_mut(payload.len() as u32) };
     dst.copy_from_slice(payload);
     b.into_raw()
 }
